backend/requirements.txt
astroid==2.3.3
autopep8==1.4.4
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
six==1.13.0
SQLAlchemy==1.3.12
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
